+++ tsml_eval/evaluation/storage/tests/test_io.py
@@ -4,7 +4,7 @@ from tsml_eval.evaluation.storage.classifier_results import ClassifierResults
 from tsml_eval.evaluation.storage.clusterer_results import ClustererResults
 from tsml_eval.evaluation.storage.forecaster_results import ForecasterResults
 from tsml_eval.evaluation.storage.regressor_results import RegressorResults
-from tsml_eval.testing.test_utils import _TEST_OUTPUT_PATH, _TEST_RESULTS_PATH
+from tsml_eval.testing.testing_utils import _TEST_OUTPUT_PATH, _TEST_RESULTS_PATH
 from tsml_eval.utils.validation import validate_results_file
