+++ tsml_eval/publications/y2023/distance_based_clustering/run_distance_experiments.py
@@ -18,7 +18,7 @@ from tsml_eval.publications.y2023.distance_based_clustering.set_distance_cluster
 from tsml_eval.publications.y2023.distance_based_clustering.tests import (
     _DISTANCE_TEST_RESULTS_PATH,
 )
-from tsml_eval.testing.test_utils import _TEST_DATA_PATH
+from tsml_eval.testing.testing_utils import _TEST_DATA_PATH
 from tsml_eval.utils.arguments import parse_args
 from tsml_eval.utils.experiments import _results_present
