+++ tsml_eval/publications/y2023/tser_archive_expansion/tests/test_run_experiments.py
@@ -7,7 +7,7 @@ from tsml_eval.publications.y2023.tser_archive_expansion import _run_experiment
 from tsml_eval.publications.y2023.tser_archive_expansion.tests import (
     _TSER_ARCHIVE_TEST_RESULTS_PATH,
 )
-from tsml_eval.testing.test_utils import _TEST_DATA_PATH
+from tsml_eval.testing.testing_utils import _TEST_DATA_PATH
 from tsml_eval.utils.tests.test_results_writing import _check_regression_file_format
